src/requirements.txt
google-genai
PyYAML
requests
shutil
pytest
tinify
convertapi
python-dotenv